+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The main API is stable, but expect some amount of breaking changes around major 
 - Add strict `Accept` header checking for `RouteSpec`s based on `producing()` settings. Option for non-strictness.
 
 ## 13.5.0 (uncut)
+- Removal of Generics in `RequestFilters`. This should make code using these filters easier.
 - Massive refactor of `Parameter` types and traits. This should be transparent to end users.
 
 ## 13.4.0